예치금으로 주문하기

+++ 2Week_Mission/src/main/java/com/ll/exam/final__2022_10_08/app/order/entity/Order.java
@@ -67,7 +67,7 @@ public class Order extends BaseEntity {
         isRefunded = true;
     }
 
-    public int getPayPrice() {
+    public long getPayPrice() {
         int payPrice = 0;
         for (OrderItem orderItem : orderItems) {
             payPrice += orderItem.getPayPrice();
